{{cookiecutter.repo_name}}/webapp/requirements_dev.txt
-r requirements.txt

fabric
ansible
django-environ
django-debug-toolbar